+++ src/libs/test/TransactionRecoveryAPI/AxelarRecoveryAPI.spec.ts
@@ -46,7 +46,43 @@ describe("AxelarDepositRecoveryAPI", () => {
           details: txDetails.gas_paid,
         },
         executed: txDetails.executed,
-        callback: undefined
+        callback: undefined,
+      });
+    });
+
+    test("it should return 'GMPStatus.DEST_EXECUTING' when the transaction is still in process", async () => {
+      const txHash = "0x123456789";
+      const txDetails = {
+        call: {
+          transactionHash: txHash,
+        },
+        gas_paid: {
+          transactionHash: txHash,
+        },
+        approved: {
+          transactionHash: txHash + "1",
+        },
+        executed: {
+          transactionHash: txHash + "2",
+        },
+        gas_status: "gas_paid",
+        status: "executing",
+      };
+
+      jest.spyOn(api, "fetchGMPTransaction").mockResolvedValueOnce(txDetails);
+
+      const status = await api.queryTransactionStatus(txHash);
+
+      expect(status).toEqual({
+        status: GMPStatus.DEST_EXECUTING,
+        error: undefined,
+        callTx: txDetails.call,
+        gasPaidInfo: {
+          status: GasPaidStatus.GAS_PAID,
+          details: txDetails.gas_paid,
+        },
+        executed: txDetails.executed,
+        callback: undefined,
       });
     });
 
@@ -82,7 +118,7 @@ describe("AxelarDepositRecoveryAPI", () => {
           details: txDetails.gas_paid,
         },
         executed: txDetails.executed,
-        callback: undefined
+        callback: undefined,
       });
     });
 
@@ -151,7 +187,7 @@ describe("AxelarDepositRecoveryAPI", () => {
           details: txDetails.gas_paid,
         },
         executed: txDetails.executed,
-        callback: undefined
+        callback: undefined,
       });
     });
 
@@ -186,7 +222,7 @@ describe("AxelarDepositRecoveryAPI", () => {
           details: txDetails.gas_paid,
         },
         executed: txDetails.executed,
-        callback: undefined
+        callback: undefined,
       });
     });