+++ test/integration/app.integration-spec.ts
@@ -1,3 +1,11 @@
-import { depositAddressSource } from "./parts/01.deposit-address.spec";
+// import { depositAddressSingle } from "./parts/00.deposit-address-single.spec";
+// import { depositAddressParallel } from "./parts/01.deposit-address.spec";
+// import { restError } from "./parts/02.rest-error.spec";
+// import { goerliLink } from "./parts/03.goerli-link.spec";
 
-depositAddressSource();
+// describe("Integration Testing", () => {
+//   depositAddressSingle();
+//   // depositAddressParallel();
+//   // restError();
+//   goerliLink();
+// });